+++ src/libs/jsx/jsx-runtime.ts
@@ -1,12 +1,4 @@
-type Props = Record<string, unknown> | null;
-
-type VNode = {
-  type: Component;
-  props: Props;
-  key: string | null;
-};
-
-type Component = string | ((props: Props) => VNode) | VNode;
+import { Component, Props, VNode } from "@/libs/types";
 
 export const Fragment = 'fragement'
